New Grid Solution + fixed mouse movement
Created custom grid node called 'Strict Grid', which confines all objects within it to be restricted to a certain size. Was definitely worth learning how to create a Godot plugin regardless of whether this overlaps with an existing Godot node's usecase. Also fixed the offset when an item is grabbed, now will calculate proper offset